+++ include/vrv/options.h
@@ -78,6 +78,8 @@ enum option_HEADER { HEADER_none = 0, HEADER_auto, HEADER_encoded };
 
 enum option_LIGATURE_OBL { LIGATURE_OBL_auto = 0, LIGATURE_OBL_straight, LIGATURE_OBL_curved };
 
+enum option_MENSURAL_RESP { MENSURAL_RESP_none = 0, MENSURAL_RESP_auto, MENSURAL_RESP_selection };
+
 enum option_MULTIRESTSTYLE {
     MULTIRESTSTYLE_auto = 0,
     MULTIRESTSTYLE_default,
@@ -152,6 +154,7 @@ public:
     static const std::map<int, std::string> s_footer;
     static const std::map<int, std::string> s_header;
     static const std::map<int, std::string> s_ligatureOblique;
+    static const std::map<int, std::string> s_mensuralResponsiveness;
     static const std::map<int, std::string> s_multiRestStyle;
     static const std::map<int, std::string> s_pedalStyle;
     static const std::map<int, std::string> s_systemDivider;
@@ -867,7 +870,7 @@ public:
     OptionBool m_ligatureAsBracket;
     OptionIntMap m_ligatureOblique;
     OptionBool m_mensuralScoreUp;
-    OptionBool m_mensuralResponsiveView;
+    OptionIntMap m_mensuralResponsiveView;
     OptionBool m_mensuralToCmn;
 
     /**