+++ src/options.cpp
@@ -47,6 +47,9 @@ const std::map<int, std::string> Option::s_header
 const std::map<int, std::string> Option::s_ligatureOblique
     = { { LIGATURE_OBL_auto, "auto" }, { LIGATURE_OBL_straight, "straight" }, { LIGATURE_OBL_curved, "curved" } };
 
+const std::map<int, std::string> Option::s_mensuralResponsiveness
+    = { { MENSURAL_RESP_none, "none" }, { MENSURAL_RESP_auto, "auto" }, { MENSURAL_RESP_selection, "selection" } };
+
 const std::map<int, std::string> Option::s_multiRestStyle = { { MULTIRESTSTYLE_auto, "auto" },
     { MULTIRESTSTYLE_default, "default" }, { MULTIRESTSTYLE_block, "block" }, { MULTIRESTSTYLE_symbols, "symbols" } };
 
@@ -1873,9 +1876,9 @@ Options::Options()
     m_ligatureOblique.Init(LIGATURE_OBL_auto, &Option::s_ligatureOblique);
     this->Register(&m_ligatureOblique, "ligatureOblique", &m_mensural);
 
-    m_mensuralResponsiveView.SetInfo(
-        "Mensural reduced view", "Convert mensural content to a more responsive view reduced to the seleceted markup");
-    m_mensuralResponsiveView.Init(false);
+    m_mensuralResponsiveView.SetInfo("Mensural responsive view",
+        "Make mensural content responsive (selection discards ligatures and editorial markup)");
+    m_mensuralResponsiveView.Init(MENSURAL_RESP_auto, &Option::s_mensuralResponsiveness);
     this->Register(&m_mensuralResponsiveView, "mensuralResponsiveView", &m_mensural);
 
     m_mensuralToCmn.SetInfo("Mensural to CMN", "Convert mensural sections to CMN measure-based MEI");