+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [unreleased]
+* Option `--mensural-responsive-view` multi-valued ("auto|none|selection")
 
 ## [6.0.1] – 2026-01-29
 * Fix ossia staff size option values